+++ apps/omsorgsdager-aleneomsorg-dialog/src/app/api/api.ts
@@ -46,7 +46,7 @@ const api = {
     post: <DataType = any, ResponseType = any>(
         endpoint: ApiEndpoint,
         data: DataType,
-        headers?: RawAxiosRequestHeaders
+        headers?: RawAxiosRequestHeaders,
     ) => {
         return axios.post<ResponseType>(endpoint, data, {
             ...axiosConfig,