+++ apps/sif-ettersending/src/app/App.tsx
@@ -55,5 +55,5 @@ root.render(
                 />
             </SoknadApplication>
         </AmplitudeProvider>
-    </SifAppWrapper>
+    </SifAppWrapper>,
 );